frontend/requirements.txt
streamlit>=1.43.1
requests>=2.31.0
python-dotenv>=1.0.0